+++ src/axom/sidre/core/Group.cpp
@@ -1033,6 +1033,31 @@ void Group::destroyGroup(const std::string& path)
 /*
  *************************************************************************
  *
+ * Detach child Group with given name or path and destroy it, and destroy
+ * any data that becomes disassociated from all Views
+ *
+ *************************************************************************
+ */
+void Group::destroyGroupAndData(const std::string& path)
+{
+  std::string intpath(path);
+  bool create_groups_in_path = false;
+  Group* group = walkPath(intpath, create_groups_in_path);
+
+  if(group != nullptr)
+  {
+    Group* targetgroup = group->detachGroup(intpath);
+    if(targetgroup != nullptr)
+    {
+      targetgroup->destroyGroupSubtreeAndData();
+      delete targetgroup;
+    }
+  }
+}
+
+/*
+ *************************************************************************
+ *
  * Detach child Group with given index and destroy it.
  *
  *************************************************************************
@@ -1042,6 +1067,24 @@ void Group::destroyGroup(IndexType idx)
   Group* group = detachGroup(idx);
   if(group != nullptr)
   {
+    delete group;
+  }
+}
+
+/*
+ *************************************************************************
+ *
+ * Detach child Group with given index and destroy it, and destroy any data
+ * that becomes disassociated from all Views
+ *
+ *************************************************************************
+ */
+void Group::destroyGroupAndData(IndexType idx)
+{
+  Group* group = detachGroup(idx);
+  if(group != nullptr)
+  {
+    group->destroyGroupSubtreeAndData();
     delete group;
   }
 }
@@ -1065,6 +1108,41 @@ void Group::destroyGroups()
   }
 
   m_group_coll->removeAllItems();
+}
+
+/*
+ *************************************************************************
+ *
+ * Detach all child Groups and destroy them, and destroy any data that
+ * becomes disassociated with all Views.
+ *
+ *************************************************************************
+ */
+void Group::destroyGroupsAndData()
+{
+  IndexType gidx = getFirstValidGroupIndex();
+  while(indexIsValid(gidx))
+  {
+    destroyGroupAndData(gidx);
+
+    gidx = getNextValidGroupIndex(gidx);
+  }
+
+  m_group_coll->removeAllItems();
+}
+
+/*
+ *************************************************************************
+ *
+ * Detach all child Groups and Views and destroy them, and destroy any data
+ * that becomes disassociated with all Views.
+ *
+ *************************************************************************
+ */
+void Group::destroyGroupSubtreeAndData()
+{
+  destroyViewsAndData();
+  destroyGroupsAndData();
 }
 
 /*
@@ -1206,6 +1284,43 @@ bool Group::createNativeLayout(Node& n, const Attribute* attr) const
   }
 
   return hasSavedViews;
+}
+
+/*
+ *************************************************************************
+ *
+ * Copy Group layout to the given Conduit node, including only the metadata
+ * for Views and not the actual data the Views hold.
+ *
+ *************************************************************************
+ */
+
+void Group::createNoDataLayout(Node& n, const Attribute* attr) const
+{
+  n.set(DataType::object());
+
+  // Dump the group's views
+  for(auto& view : views())
+  {
+    // Check that the view's name is not also a child group name
+    SLIC_CHECK_MSG(m_is_list || !hasChildGroup(view.getName()),
+                   SIDRE_GROUP_LOG_PREPEND
+                     << "'" << view.getName()
+                     << "' is the name of both a group and a view.");
+
+    if(attr == nullptr || view.hasAttributeValue(attr))
+    {
+      conduit::Node& child_node = m_is_list ? n.append() : n[view.getName()];
+      view.copyMetadataToNode(child_node);
+    }
+  }
+
+  // Recursively dump the child groups
+  for(auto& group : groups())
+  {
+    conduit::Node& child_node = m_is_list ? n.append() : n[group.getName()];
+    group.createNoDataLayout(child_node, attr);
+  }
 }
 
 /*
@@ -1490,6 +1605,14 @@ void Group::save(const std::string& path,
     n["sidre_group_name"] = m_name;
     conduit::relay::io::save(n, path, "json");
   }
+  else if(protocol == "sidre_layout_json")
+  {
+    Node n;
+    exportWithoutBufferData(n["sidre"], attr);
+    ds->saveAttributeLayout(n["sidre/attribute"]);
+    n["sidre_group_name"] = m_name;
+    conduit::relay::io::save(n, path, "conduit_json");
+  }
   else if(protocol == "conduit_hdf5")
   {
     Node n;
@@ -1504,6 +1627,13 @@ void Group::save(const std::string& path,
     createNativeLayout(n, attr);
     n["sidre_group_name"] = m_name;
     conduit::relay::io::save(n, path, protocol);
+  }
+  else if(protocol == "conduit_layout_json")
+  {
+    Node n;
+    createNoDataLayout(n, attr);
+    n["sidre_group_name"] = m_name;
+    conduit::relay::io::save(n, path, "conduit_json");
   }
   else
   {
@@ -1975,7 +2105,9 @@ Group* Group::detachGroup(IndexType idx)
  *
  *************************************************************************
  */
-bool Group::exportTo(conduit::Node& result, const Attribute* attr) const
+bool Group::exportTo(conduit::Node& result,
+                     const Attribute* attr,
+                     bool export_buffer) const
 {
   result.set(DataType::object());
   // TODO - This implementation will change in the future.  We want to write
@@ -2007,7 +2139,14 @@ bool Group::exportTo(conduit::Node& result, const Attribute* attr) const
       std::ostringstream oss;
       oss << "buffer_id_" << *s_it;
       Node& n_buffer = bnode.fetch(oss.str());
-      getDataStore()->getBuffer(*s_it)->exportTo(n_buffer);
+      if(export_buffer)
+      {
+        getDataStore()->getBuffer(*s_it)->exportTo(n_buffer);
+      }
+      else
+      {
+        getDataStore()->getBuffer(*s_it)->exportMetadata(n_buffer);
+      }
     }
   }
 
@@ -2075,6 +2214,21 @@ bool Group::exportTo(conduit::Node& result,
   }
 
   return hasSavedViews;
+}
+
+/*
+ *************************************************************************
+ *
+ * Exports the contents of the Group, excluding the data arrays held
+ * by any Buffers attached to the Views.
+ *
+ *************************************************************************
+ */
+
+bool Group::exportWithoutBufferData(conduit::Node& result,
+                                    const Attribute* attr) const
+{
+  return exportTo(result, attr, false);
 }
 
 /*